Refactor model call pipeline
